requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.3.0
Werkzeug==2.0.1
gunicorn==20.1.0
eventlet==0.33.3
requests==2.31.0
simple-websocket==0.10.1
